% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_converter.R
\name{optdf}
\alias{optdf}
\title{Optimal type selector for a data.frame}
\usage{
optdf(df)
}
\arguments{
\item{df}{data.frame to be converted}
}
\description{
Optimal type selector for a data.frame
}
\examples{
data("iris")
optdf(iris)
}
